docs: require GEMINI_API_KEY for examples
