+++ examples/analyse/README.md
@@ -2,6 +2,7 @@
 
 - **Purpose:** Analyze a document attachment with a role-specific question.
 - **Key PMFS methods:** `Attachment.Analyse`, `SetBaseDir`
+- **Requires:** `GEMINI_API_KEY`
 - **Run:**
   ```bash
   go run ./examples/analyse
