+++ examples/gates/README.md
@@ -2,6 +2,7 @@
 
 - **Purpose:** Evaluate a requirement against PMFS quality gates.
 - **Key PMFS methods:** `Requirement.EvaluateGates`
+- **Requires:** `GEMINI_API_KEY`
 - **Run:**
   ```bash
   go run ./examples/gates
