+++ examples/gemini/README.md
@@ -2,6 +2,7 @@
 
 - **Purpose:** Use the Gemini client to analyze a document and answer questions.
 - **Key PMFS methods:** `gemini.AnalyzeAttachment`, `gemini.Ask`
+- **Requires:** `GEMINI_API_KEY`
 - **Run:**
   ```bash
   go run ./examples/gemini
