+++ examples/gemini/main.go
@@ -7,22 +7,10 @@ import (
 	gemini "github.com/rjboer/PMFS/pmfs/llm/gemini"
 )
 
-// This example demonstrates using the Gemini client to analyze a document
-// and answer a free-form question. It swaps in a stub client so the example
-// runs without calling the real API. For production, remove the SetClient
-// block and ensure GEMINI_API_KEY is set; the default client will then invoke
-// Gemini directly.
+// This example demonstrates using the Gemini client to analyze a document and
+// answer a free-form question. Requires the GEMINI_API_KEY environment
+// variable.
 func main() {
-	prev := gemini.SetClient(gemini.ClientFunc{
-		AnalyzeAttachmentFunc: func(path string) ([]gemini.Requirement, error) {
-			return []gemini.Requirement{{ID: 1, Name: "Sample", Description: "From " + path}}, nil
-		},
-		AskFunc: func(prompt string) (string, error) {
-			return "stubbed answer", nil
-		},
-	})
-	defer gemini.SetClient(prev)
-
 	reqs, err := gemini.AnalyzeAttachment("testdata/spec1.txt")
 	if err != nil {
 		log.Fatalf("analyze: %v", err)
