+++ examples/integration/README.md
@@ -2,6 +2,7 @@
 
 - **Purpose:** Demonstrate an end-to-end flow: analyze a document, store a requirement, query a role, and evaluate gates.
 - **Key PMFS methods:** `gemini.AnalyzeAttachment`, `Requirement.Analyse`, `Requirement.EvaluateGates`
+- **Requires:** `GEMINI_API_KEY`
 - **Run:**
   ```bash
   go run ./examples/integration
