+++ examples/mock/README.md
@@ -1,0 +1,10 @@
+# Mock Example
+
+- **Purpose:** Demonstrate stubbing the Gemini client.
+- **Key PMFS methods:** `gemini.SetClient`
+- **Requires:** none (uses a stubbed client)
+- **Run:**
+  ```bash
+  go run ./examples/mock
+  ```
+
