+++ examples/project/README.md
@@ -1,0 +1,10 @@
+# Project Example
+
+- **Purpose:** Demonstrate project setup, attachment ingestion, requirement analysis, and gate evaluation.
+- **Key PMFS methods:** `EnsureLayout`, `Attachment.Analyze`, `Requirement.Analyse`, `Requirement.EvaluateGates`
+- **Requires:** `GEMINI_API_KEY`
+- **Run:**
+  ```bash
+  go run ./examples/project
+  ```
+
